# docs/swagger.yaml
basePath: /api
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  internal_domain_demo.Demo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for a demo API.
  title: Demo API
  version: "1.0"
paths:
  /demos/:
    get:
      description: Retrieve all demo entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_domain_demo.Demo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all Demos
      tags:
      - demos
    post:
      consumes:
      - application/json
      description: Create a new demo entry
      parameters:
      - description: Demo to create
        in: body
        name: demo
        required: true
        schema:
          $ref: '#/definitions/internal_domain_demo.Demo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_domain_demo.Demo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create a new Demo
      tags:
      - demos
  /demos/{id}:
    delete:
      description: Delete a demo entry by ID
      parameters:
      - description: Demo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a Demo
      tags:
      - demos
    get:
      description: Get details of a demo entry by ID
      parameters:
      - description: Demo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_domain_demo.Demo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a Demo by ID
      tags:
      - demos
    put:
      consumes:
      - application/json
      description: Update a demo entry by ID
      parameters:
      - description: Demo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated demo
        in: body
        name: demo
        required: true
        schema:
          $ref: '#/definitions/internal_domain_demo.Demo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_domain_demo.Demo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update an existing Demo
      tags:
      - demos
schemes:
- http
swagger: "2.0"